fringematrix5-5loh: extract shared e2e helpers
Move duplicated helpers into e2e/helpers/ modules:

- e2e/helpers/wireframe.ts: waitForLoaderToFinish, waitForWireframeVisible,
  waitForWireframeHidden — bodies identical across previous copies.
- e2e/helpers/author-browse.ts: ZORT_HANDLE, ZORT_HASH, gotoZortAuthorDetail.
  Helper now owns the test.skip when the fixture returns 0 images (no
  BLOB_READ_WRITE_TOKEN) so callers don't have to remember the gate.

Refactor in-scope specs to import from the shared modules:

- e2e/author-browse-mode.spec.ts — drop local defs + per-test
  if (total === 0) gates (helper handles it now).
- e2e/lightbox-animations.spec.ts — drop local wireframe + loader helpers.
- e2e/lightbox-author-row.spec.ts — drop local waitForLoaderToFinish copy.

Other specs that also duplicate waitForLoaderToFinish (lightbox-outside-click,
lightbox-panel-animation, lightbox-redesign, modal-focus-trap) are left
untouched per the bead's "no other spec refactors" out-of-scope clause.

Behavior identical; npm test passes (635 tests), playwright --list
parses all 84 e2e tests.
