+++ e2e/author-browse-mode.spec.ts
@@ -1,4 +1,5 @@
 import { test, expect, Page } from '@playwright/test';
+import { ZORT_HANDLE, gotoZortAuthorDetail } from './helpers/author-browse';
 
 /**
  * E2E coverage for author-browse mode and the info-box in-app links that ride
@@ -35,42 +36,6 @@ import { test, expect, Page } from '@playwright/test';
  * in that case rather than failing.
  */
 
-const ZORT_HANDLE = '@Zort70';
-const ZORT_HASH = `#authors/${encodeURIComponent(ZORT_HANDLE)}`;
-
-async function waitForLoaderToFinish(page: Page) {
-  const loader = page.getByRole('dialog', { name: 'Loading' });
-  if (await loader.isVisible().catch(() => false)) {
-    await loader.waitFor({ state: 'detached' });
-  }
-}
-
-/**
- * Navigate to the @Zort70 author detail page and wait for it to populate.
- * Returns the number of thumbnails in the grid (== the count the lightbox
- * counter should reflect). When the environment has no blob token the grid
- * comes back empty; the caller is expected to skip in that case.
- */
-async function gotoZortAuthorDetail(page: Page): Promise<number> {
-  await page.goto(`/${ZORT_HASH}`);
-  await waitForLoaderToFinish(page);
-
-  // The grid only renders once the /api/authors/@Zort70 fetch resolves with
-  // images. In empty-blob environments the page shows the "no images" status
-  // and the grid never appears — wait with a short escape hatch by polling
-  // the count directly.
-  const grid = page.locator('[data-testid="author-images-grid"]');
-  await expect.poll(
-    async () => (await grid.count()) > 0,
-    { timeout: 15_000 }
-  ).toBe(true).catch(() => {
-    // fall through — count will read 0 and the caller will skip
-  });
-
-  const cards = grid.locator('.card img');
-  return cards.count();
-}
-
 /**
  * Read the lightbox HUD's "N OF M" counter as a {current, total} pair.
  * Asserts the HUD is in the expected shape so a regression in the HUD
@@ -102,9 +67,6 @@ async function readEpisodeName(page: Page): Promise<string> {
 test.describe('Author-browse mode — lightbox navigation', () => {
   test('clicking a thumbnail opens the lightbox with the author total as the counter', async ({ page }) => {
     const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
 
     const grid = page.locator('[data-testid="author-images-grid"]');
     await grid.locator('.card img').first().click();
@@ -119,9 +81,6 @@ test.describe('Author-browse mode — lightbox navigation', () => {
 
   test('ArrowRight cycles through the author list and the IMAGE DETAILS panel updates across campaigns', async ({ page }) => {
     const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
     // We need to see at least one campaign boundary to verify the
     // per-image campaign resolution (pyd). @Zort70 is known to span
     // multiple campaigns, but if the fixture ever shrinks to a single
@@ -166,9 +125,6 @@ test.describe('Author-browse mode — lightbox navigation', () => {
 
   test('NEXT button in the toolbar advances and PREVIOUS reverses', async ({ page }) => {
     const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
     if (total < 2) test.skip(true, 'Need at least 2 images');
 
     const grid = page.locator('[data-testid="author-images-grid"]');
@@ -187,10 +143,7 @@ test.describe('Author-browse mode — lightbox navigation', () => {
   });
 
   test('closing the lightbox returns to the author detail page (hash unchanged)', async ({ page }) => {
-    const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
+    await gotoZortAuthorDetail(page);
 
     const grid = page.locator('[data-testid="author-images-grid"]');
     await grid.locator('.card img').first().click();
@@ -208,10 +161,7 @@ test.describe('Author-browse mode — lightbox navigation', () => {
 
 test.describe('Author-browse mode — info-box in-app links', () => {
   test('clicking the AUTHOR handle inside the lightbox navigates to the author gallery', async ({ page }) => {
-    const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
+    await gotoZortAuthorDetail(page);
 
     const grid = page.locator('[data-testid="author-images-grid"]');
     await grid.locator('.card img').first().click();
@@ -239,10 +189,7 @@ test.describe('Author-browse mode — info-box in-app links', () => {
   });
 
   test('clicking EPISODE NAME inside the lightbox switches the gallery to that campaign', async ({ page }) => {
-    const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
+    await gotoZortAuthorDetail(page);
 
     const grid = page.locator('[data-testid="author-images-grid"]');
     await grid.locator('.card img').first().click();
@@ -293,10 +240,7 @@ test.describe('Author-browse mode — info-box in-app links', () => {
   });
 
   test('clicking HASHTAG inside the lightbox switches the gallery to that campaign', async ({ page }) => {
-    const total = await gotoZortAuthorDetail(page);
-    if (total === 0) {
-      test.skip(true, 'No images returned for @Zort70 (BLOB_READ_WRITE_TOKEN likely missing).');
-    }
+    await gotoZortAuthorDetail(page);
 
     const grid = page.locator('[data-testid="author-images-grid"]');
     await grid.locator('.card img').first().click();
